+++ packages/plugins/pointer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @splicetree/plugin-pointer
 
+## 1.1.0
+
 ## 1.0.0
 
 ### Major Changes